+++ src/js/content/donations.js
@@ -1,5 +1,6 @@
 content.donations = (() => {
-  let amount = 0
+  let amount = 0,
+    collected = 0
 
   return {
     add: function (value = 0) {
@@ -8,22 +9,34 @@ content.donations = (() => {
       return this
     },
     amount: () => amount,
+    collect: function (value = 0) {
+      value = engine.fn.clamp(Math.abs(value) || 0, 0, amount)
+
+      amount -= value
+      collected += value
+
+      return this
+    },
+    collected: () => collected,
     export: () => ({
       amount,
+      collected,
     }),
     has: (value = 1) => value <= amount,
     import: function (data = {}) {
       amount = data.amount || 0
+      collected = data.collected || 0
 
       return this
     },
     remove: function (value = 0) {
-      amount -= Math.abs(value) || 0
+      amount -= engine.fn.clamp(Math.abs(value) || 0, 0, amount)
 
       return this
     },
     reset: function () {
       amount = 0
+      collected = 0
 
       return this
     },